Moved test app into subdirectory
- Gives a nice separation between test app and test configuration
- Allows us to cleanly add separate test apps
