+++ wagtail/contrib/wagtailroutablepage/tests.py
@@ -1,7 +1,7 @@
 from django.test import TestCase, RequestFactory
 
 from wagtail.wagtailcore.models import Page, Site
-from wagtail.tests.models import RoutablePageTest, routable_page_external_view
+from wagtail.tests.testapp.models import RoutablePageTest, routable_page_external_view
 from wagtail.contrib.wagtailroutablepage.templatetags.wagtailroutablepage_tags import routablepageurl
