+++ wagtail/tests/testapp/wagtail_hooks.py
