+++ wagtail/wagtailcore/tests/tests.py
@@ -7,7 +7,7 @@ from django.utils.safestring import SafeString
 from wagtail.wagtailcore.models import Page, Site
 from wagtail.wagtailcore.templatetags.wagtailcore_tags import richtext
 from wagtail.wagtailcore.utils import resolve_model_string
-from wagtail.tests.models import SimplePage
+from wagtail.tests.testapp.models import SimplePage
 
 
 class TestPageUrlTags(TestCase):
