+++ wagtail/wagtailsearch/tests/test_indexed_class.py
@@ -3,7 +3,7 @@ import warnings
 from django.test import TestCase
 
 from wagtail.wagtailsearch import index
-from wagtail.tests import models
+from wagtail.tests.testapp import models
 from wagtail.tests.utils import WagtailTestUtils
